requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
google-generativeai==0.2.0
gunicorn==20.1.0